refactor, moving common code into util
