+++ iotd.py
@@ -1,110 +1,26 @@
 import os
-import subprocess
-
-import pandas as pd
 import requests
-
-from requests_html import HTML, HTMLSession
 from bs4 import BeautifulSoup
+from util import get_feed, download_image
+
+iotd_url = "https://www.nasa.gov/rss/dyn/lg_image_of_the_day.rss"
+target_folder = "./images/iotd"
 
 
-def get_source(url):
-    """Return the source code for the provided URL. 
-
-    Args: 
-        url (string): URL of the page to scrape.
-
-    Returns:
-        response (object): HTTP response object from requests_html. 
-    """
-    session = HTMLSession()
-    response = session.get(url)
-    return response
+def get_iotd_image_url(post_url):
+    last_post = requests.get(post_url)
+    data = str(last_post.content)
+    soup = BeautifulSoup(data)
+    image = soup.find("meta", property="og:image")
+    image_url = image.attrs.get("content", "")
+    return image_url
 
 
-def get_feed_pd(url):
-    """Return a Pandas dataframe containing the RSS feed contents.
-
-    Args: 
-        url (string): URL of the RSS feed to read.
-
-    Returns:
-        df (dataframe): Pandas dataframe containing the RSS feed contents.
-    """
-    
-    response = get_source(url)
-    
-    df = pd.DataFrame(columns = ['title', 'link', 'pubDate', 'guid', 'description'])
-
-    with response as r:
-        items = r.html.find("item", first=False)
-        
-        for item in items:        
-
-            title = item.find('title', first=True).text
-            link = item.find('link', first=True).text
-            pubDate = item.find('pubDate', first=True).text
-            guid = item.find('guid', first=True).text
-            description = item.find('description', first=True).text
-
-            row = {'title': title, 'link': link, 'pubDate': pubDate, 'guid': guid, 'description': description}
-            df = df.append(row, ignore_index=True)
-
-    return df
-
-
-def get_feed(url, item_count: int):
-    """Return a Pandas dataframe containing the RSS feed contents.
-
-    Args: 
-        url (string): URL of the RSS feed to read.
-
-    Returns:
-        df (dataframe): Pandas dataframe containing the RSS feed contents.
-    """
-    
-    response = get_source(url)
-    fields = {"title": "text", "link": "html", "pubDate": "text", "guid": "text", "description": "text"}
-    
-    feed_items = []
-    with response as r:
-        response_items = r.html.find("item", first=False)
-        if item_count: 
-            response_items = response_items[:item_count]
-
-        for response_item in response_items:
-            feed_item = dict((k, getattr(response_item.find(k, first=True), v)) for k, v in fields.items())
-            feed_item["link"] = feed_item["link"].replace("<link/>", "")
-            feed_items.append(feed_item)
-
-    return feed_items
-
-
-url = "https://www.nasa.gov/rss/dyn/lg_image_of_the_day.rss"
-last_post_url = get_feed(url, 1)[0]["link"]
-
-last_post = requests.get(last_post_url)
-
-data = str(last_post.content)
-soup = BeautifulSoup(data)
-mtea = soup.find("meta")
-image = soup.find("meta", property="og:image")
-image_url = image.attrs.get("content", "")
-
-print(image_url)
-
-if image_url:
-    image_name = image_url.split("/")[-1]
-
-    target_folder = "./images/iotd"
+if __name__ == '__main__':
     os.makedirs(target_folder, exist_ok=True)
 
-    target_path = os.path.join(target_folder, image_name)
-    if os.path.exists(target_path):
-        print(f"target file {target_path} already exists")
-        os.remove(target_path)
-    else:
-        print(f"{image_url} ==> {target_path}")
-        import urllib.request
-        urllib.request.urlretrieve(image_url, target_path)
-        subprocess.call(('open', target_path))
+    last_post_url = get_feed(iotd_url, 1)[0]["link"]
+    image_url = get_iotd_image_url(last_post_url)
+
+    if image_url:
+        download_image(image_url, target_folder)
